+++ pkgs/test_core/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.4.14-dev
 
+* Fix `spawnHybridUri` support for `package:` uris.
+
 ## 0.4.13
 
 * Re-publish changes from 0.4.12.